Fix remote autoselection when not on a branch
We currently have code that determines which remote to us automatically.
For example, if there is only one remote, we want to use it, no matter
what it's named, because there's no other logical choice.

However, if we were not on any branch, we'd always end up picking
"origin", even if that remote didn't exist. This appears to have been
because we wanted to avoid a lookup with an empty branch name in the
configuration. Doing so is harmless, though, so let's simply check that
the branch name isn't empty before we execute that condition, and fall
back to the rest of the checks if it is. This ensures that we always
pick the only remote when there's just one.


Former-commit-id: f6f4f6a8e4f6f5713b3325202a5709d437d92617