Switch CI to use GitHub Actions
Currently, we have three different CI systems that handle our CI: Travis
for Linux, CircleCI for macOS, and AppVeyor for Windows.  This results
in widely varying performance across systems and the need to maintain
code that works differently across different CI systems.

In addition, we'd like to use GitHub Actions to automate the release
process, so it makes sense to use it for CI as well.  Switch over by
adding a CI workflow that runs our existing jobs.  Ensure that we filter
out the environment variables that GitHub Actions provides, since they
will cause tests that run "git lfs env" to fail.

Add a script for those jobs where we build a custom Git and install the
appropriate dependencies.  In the cibuild script, hoist the Windows
handling to the top, set a specific environment variable for us to
remember that we're on Windows, and then disable locking, which fails on
Windows and causes the testsuite to abort.  These same environment
variables were set for AppVeyor and are also needed on Windows
development systems.


Former-commit-id: c32820806229c3f42364d989f7a8597f73cb107b