t: fix flaky tests due to race condition
It's possible that the final "Downloading objects" message might not be
printed in some cases due to a race condition.  (The corresponding
upload message, on the other hand, will always be printed because it
runs before the actual push.)

To avoid spurious failures, don't look for this message in several tests
where this race condition commonly occurs.  Instead, use "git lfs fsck"
to verify that we downloaded all the objects we expected to get.

In addition, several of the tests also check that the expected number of
objects is correct, which provides additional confidence that our test
did in fact succeed.


Former-commit-id: cd754792c02eea1a3ac732a623686afa4e2d1640