t: add a test helper equivalent to realpath(1)
On Actions, our Windows build root has a path with a short name.  This
causes significant grief, since our tests expect one version of the path
and we output another, causing our tests for equality to fail.  Since
Windows lacks a realpath(1) utility, provide a small wrapper around the
Go primitives to provide a limited version of this utility for use only
in our tests.


Former-commit-id: 391689b02e80bef335112cfe1a5517a06787a684