t: add a test helper to print the canonical path
There are several places in our testsuite where we want to produce a
canonical path.  For example, in t/t-env.sh, we compare the expected
output, which we synthesize, with the output from the git-lfs binary
we're testing.  The binary will have canonicalized the path, and
therefore so should we.  This is especially noticeable on Windows on
Actions, since cygpath -w produces a noncanonical path.

Introduce a test helper to produce a canonical native path that combines
the functionality of our existing native_path helper and the
lfstest-realpath binary we introduced in an earlier commit.  Add a form
that handles escaped paths as well.


Former-commit-id: e0e8cdcbf127444505ed69f4879143de9609cf6f