commands/status: canonicalize all paths
When we're trying to make a path in the repository relative to the
current working directory, the path to the repository that we use will
already have been canonicalized.  Ensure that the path to the current
working directory is also canonicalized, so that we don't end up
creating a long chain of ".." components when the repository path and
the working directory differ in canonicalization.


Former-commit-id: 56fc55a33f5218421fc5cd7bcb1bccd98c9fdc90