script/upload: handle upload of unsigned hashes
When we automate the release process using GitHub Actions, we'll want to
upload an unsigned list of hashes that the maintainer can then sign and
re-upload.  In order to do so, we'll need to handle the case where we're
uploading an unsigned list of hashes, so teach our upload script about
plain files named "sha256sums".


Former-commit-id: bca33bfe745e8f55cb30709641bb399b598405fb