script/upload: add a --skip-verify flag
While the ability to verify assets is important to ensuring our releases
are intact, we also want to be able to skip this step if we're running
in GitHub Actions.  We may upload assets in multiple steps and not want
to verify immediately, or even have a signed manifest until the final,
manual stages of the process.  Add a --skip-verify flag that allows
skipping the verification process.


Former-commit-id: 119a5f184d7bdc1cd9ea556ba64f89c6bace7f2e