t: add a test for shell characters in paths
Now that we've included gitobj v1.4.1, add a test for the failure case
we had in the past to prevent us from regressing this behavior.  Check
that shell characters within a pathname don't cause a failure when we
open the pack file.  Choose brackets since these are available on all
platforms, including Windows.


Former-commit-id: c384de075cc4cb01023b1c74dccd5f64da2e46e0