tools: require cgo for Darwin CloneFile
On Darwin, we're importing "C" in the Darwin-specific file, so flag it
as requiring cgo and specify the generic file for when we're not using
cgo.  This ensures that cross-compilation continues to work without
errors.  Since this is just a performance optimization, there's
relatively little problem with not providing the CloneFile function in a
non-cgo environment.


Former-commit-id: 08d9e836a505d5bdca33b5293a3c1776fb2c55b7