ci: force Windows Git version to 2.22.0
In Git for Windows 2.23, the /dev and /dev/fd directories did not get
created.  As a result, it's not possible to use a process substitution
in bash, which breaks a significant portion of the test suite.  In the
mean time, let's downgrade Git for Windows to 2.22.0, which is what
Actions used to use and worked for us.  We can remove this hack when the
next Git for Windows release occurs.


Former-commit-id: 52257344da89a7ba0891d49ac4d60a8e04563aad