docker: run interactively only if we have a TTY
When building packages in containers, we pass options to docker.
However, when running in a noninteractive environment such as GitHub
Actions, we can't use the -i option, since standard input will not be a
TTY and docker will consequently fail.  Check for a TTY and pass the -i
option only if one is present.


Former-commit-id: 1a5eaa6f47c937ba88bc4314f9c6ab83c65d5754