Makefile: make gzipped tarball really gzipped
When we create a tarball with Windows assets, we create it with the
extension ".tar.gz" and we later on extract it with the -z flag to tar.
However, we fail to actually compress the tarball itself.  This causes
no problems with macOS's tar, which automatically handles compression
based on magic numbers, but it breaks with GNU tar, which we use during
the Windows build process.  Ensure that we actually compress the tarball
so that we can decompress it in a later stage.


Former-commit-id: 97e26b9c9b0433ac7c82077f54c4ede87e0db421