Accept all local references with git lfs push
This change makes it possible to run git lfs push with any local
reference, regardless of whether it’s stored in the conventional places
(refs/heads and refs/tags) or not. This is to bring the behavior of git
lfs push in line with git push, which doesn’t limit source references in
refspecs to the refs/heads and refs/tags scopes either.


Former-commit-id: ac2abfadb5e0e3f5621006c8a6817dc96c7ef46b