workflows/ci: build Windows binaries during CI
When proposing a PR that may fix a user issue, it's helpful for folks to
be able to test the fix easily.  For most macOS and Linux users, a Go
compiler and the other relevant tools are easily available via a package
manager, but most Windows users don't have a common package manager
available.

Let's build Windows binaries as part of the CI process and upload them
as artifacts for users to consume for testing purposes.


Former-commit-id: 685485eef853de1992255366e77bf27c528b29ec