workflows/ci: build Unix assets
Even though most Mac and Linux users have a ready set of build tools
available, it's still tricky to explain to users how to build Git LFS,
since they may or may not be familiar with Go or makefiles.  Let's build
a set of binaries for each of these platforms as well and include them
as artifacts for easy testing.


Former-commit-id: 84e1b6acc60f7a72423130462c042e49575bb01a