t: fix flaky tests in CI
When we download objects, it's possible that one or more messages may
not be printed because the transfer meter is paused.  This is not
normally a problem for production use, since we'll print a later update
out, but in our tests, we check for specific success messages, which
leads to flaky tests when those exact messages happen to be skipped.
Since our CI hosts are very fast and good at triggering race conditions,
we see this frequently there.

Instead of checking for these exact messages, let's instead check that
the objects we want to have downloaded have in fact been downloaded and
that git lfs fsck passes, which are more reliable indicators of a
successful download.  In many cases, we're already doing one or more of
these, so we already have confidence that everything's working
correctly.


Former-commit-id: 65e8a8b8cb88d0cf4fa6701a109d0a67bc7bd332