Honor lfs.url when deciding on transfer adapters
Currently, if you have a remote with a file URL but have lfs.url set to
point to a non-file URL, your transfers will fail.  This occurs because
we look up whether to use a transfer adapter based on the remote URL,
but we also look up whether to default to the built-in transfer adapter
based on the remote URL, not the endpoint URL.  We then attempt to pass
the HTTP URL to the standalone transfer adapter, which doesn't work.

Look up the endpoint URL as well and pass that to the code which
determines whether we should use the builtin transfer adapter.  If the
actual endpoint is not a file URL, then we won't suggest the builtin
transfer adapter.


Former-commit-id: 28f9b73d2f2bc529ce4385d372ab900e3763b854