tools: handle missing cygpath gracefully
It looks like there are situations in which some Git for Windows
installations don't have the cygpath binary.  In that case, we end up
picking an empty string, which leads us to install hooks and LFS objects
in the current directory.  Fix this by trying cygpath, and if it fails
due to a path lookup error, returning the original path as we got it.


Former-commit-id: 354c8a53d92581a1b24310cf07e7612ee8b4f81b