docs: explain shell metacharacters better
It's a surprise to many users that bracket characters are treated
specially in the shell, especially Windows users or other users who are
not intimately familiar with the various peculiarities of shell.  As a
consequence, users are also surprised when putting brackets into a `git
lfs track` pattern doesn't work as expected.

Let's help these users out by explaining exactly what characters are
special and what are ignored, since folks don't always dive into the Git
attribute documentation.  Let's also provide an example of tracking a
file with brackets in the name to reinforce the use of the `--filename`
option as well.


Former-commit-id: 24ba71af7ee3bfadc854b1a96b58b93cba3cf80a