Interpret relative hook paths as relative to working tree
Git specifies that relative hook paths are specified relative to the
working tree in a non-bare repository.  However, we were interpreting it
relative to the working directory, which meant that we could install
hooks in an undesired place if the user was in a subdirectory of the
repository.  Fix this by installing relative to the root of the working
tree as Git does.


Former-commit-id: 639a9552de7a052aa3d09a7baa8e10dd943fbb0f