Don't abort with newer Git when in a bare repo
In Git 2.25, git rev-parse --show-toplevel will fail when not in a work
tree.  This currently causes our test suite to abort because there are
myriad places that we use bare repositories, in addition to the user
impact.  Fix this by adjusting our code to fall back to querying just
the Git directory if we get an exit code of 128 from Git, which
indicates that Git called die().

The alternative would be to inspect the error message, which should
theoretically be stable, since it comes from die() (it will at least not
be localized), but that is potentially brittle.  We can revisit this
decision in a few Git releases if it remains stable.


Former-commit-id: 451a8bb4a2581d6e0592795a7bd0d1c424699e2f