install: don't print error if run outside repository
As of 2.9.2, if run outside of a repository, Git LFS will produce an
error if `git lfs install` is run outside of a repository.  This occurs
because we look for a specific string in the error message to detect if
the directory is not a repository, but that message was no longer
included since the standard output of `git rev-parse` is not included.

Include the standard error output in the message so that we correctly
detect whether the directory is a repository or not and avoid printing
an error message in this case if it is not.


Former-commit-id: f68ce7517bc6f672a1419b5c70f67792a5c281c6