lfshttp: improve proxy support
Currently, our proxy support has some limitations.  Notably, we don't
handle wildcards in the no_proxy environment variable.

Unfortunately for us, there's no standard as to how these environment
variables are supposed to be handled, so any attempt we made to handle
this ourselves would likely be incomplete.  However, fortunately for us,
our needs are simple: we need standard behavior except that a user can
use a proxy for localhost, and we need to be able to read from a source
other than the environment.

The solution is to use the Go extension httpproxy module along with a
little bit of custom configuration and URL rewriting, which means we
don't have to worry about maintaining the complexity of parsing proxy
support.  This also means we can drop a large amount of complex (and
subtly wrong) code.


Former-commit-id: acd77c36f9300babddedc2567dff669ade1723e4