.github/workflows: temporarily fix path on Windows
There's a bug with Actions right now that causes the path to be set
incorrectly, and as a result, this causes Windows binaries to be
executed in some cases instead of the binaries that come with Git Bash.
This shows up in our test suite as a hang during the execution of
t-credentials.sh.

Since we can't build or release software while the Actions workflows
aren't working, let's adjust the path temporarily so that things work as
expected.


Former-commit-id: 73099e6493b444bb463c08c9e1d5991afb255f2d