track: detect duplicate patterns with --filename
When we add a pattern with git lfs track, we inform the user if the
pattern is already supported.  However, we fail to do so when the
--filename argument is passed, due to not having encoded the pattern at
that point.

To solve this problem, hoist the code that encodes the filename as a
pattern so it happens before we check if we already have such a pattern.


Former-commit-id: ab58f576bf176cf77638ecda51150260e5f12875