Add an option to control warning about files larger than 4 GiB
Git on Windows is currently incapable of handling files larger than 4
GiB in size.  This is because Git uses "unsigned long" for numerous
types of variables representing files, which on a Unix system is an
appropriate way to specify a pointer-sized integer, since Unix systems
are either ILP32 or LP64.  However, Windows is LLP64, meaning that a
variable of type unsigned long is always 32 bits.

We warn about this case to help alert users to potential corruption when
they're working across platforms.  However, there are some users and
projects who don't use Windows and therefore don't have this problem.
For those users and projects, this warning is not helpful.  Let's add
an option to control that warning so folks who don't wish to see it can
disable it.


Former-commit-id: 87fb0fff815698bc2a66b475e5a1042dcee5e4e7