Check error when creating local storage directory
When we have an error creating the local storage directory, such as if
the permissions are wrong, we fail to produce a helpful error message,
since we fail to check the error we produce.

Let's take the error we get in such a case and pass it through to the
transfer queue, where we can handle it as any other error.  This means
that if only one directory has a problem, we can transfer the rest of
the objects and fail for only the one problematic object.


Former-commit-id: a514c7cf435760c52badf53010733957e496327e