tools: remove the use of cgo on Linux
We currently import the C package solely for a single value: the FICLONE
ioctl, which is a constant.  Since this is a constant and also part of
the Linux kernel ABI, there's no reason we need to have the C compiler
generate it for us every time we build.  We can instead hard-code the
constant in place, since it will never change.

This avoids the need for cgo in order to have a functional invocation of
git lfs dedup on Linux.


Former-commit-id: a27bda0eb938c1f0b33efa1d3e9881992a6a7069