commands: support git lfs help config invocation
Because "config" is not a command, Cobra returns an error when
the user invokes "git lfs help config" on the command line,
so we just print an error message and the generic usage
instead of returning the same content as one gets from
"man git-lfs-config" for git-lfs-config(5).

This differs from Git's own support for running, for instance,
"git help gitrepository-layout" for gitrepository-layout(5)
or "git help githooks" for githooks(5).

Therefore we specifically check if the supplied argument
to the "help" command is "config" and in that case look up
a dummy argument so we can proceed.


Former-commit-id: 590881e54cdb8d59657b546f46e89c1b290198aa