lfshttp: allow renegotiation
There are some TLS servers that want to perform a renegotiation to
request users use a client certificate.  Allow us to use renegotiation
as a client when making a TLS connection.  The DoS potential for this is
low, since the server must not only perform another key exchange, but
must re-sign the request, leading to a larger performance impact on the
server than the client.


Former-commit-id: 4e9b2ba96f70e8e8231a9bb5e6d66c148a846ccf