migrate import: set text to unspecified for excluded fields
When we specify the --exclude option to git lfs migrate import to
exclude certain patterns, we currently specify the files as text.
However, in many cases, the files are not text, so it doesn't make sense
to set them that way.

Let's set the text attribute to unspecified instead and let Git
automatically figure out whether the file should be treated as text or
binary.  That's at least less likely to result in silent breakage and
will likely be less surprising to the user.


Former-commit-id: c8524d256397abc44f75415ff8c73b466515afda