t: ensure local binary always used in tests
When running our test suite, including the Go tests, we want to ensure
that our locally compiled "git-lfs" binary is the one which is used
whenever "git-lfs" is invoked, for instance as a Git filter command
or as the transfer queue's default custom adapter.  Otherwise we may
see unexpected test failures or successes, if another "git-lfs" binary
installed elsewhere on the system is used instead.

Therefore we define an init() function in the common test utilities
package which prepends to the PATH environment variable an absolute path
to the working tree's "bin/" directory, to ensure our "git-lfs" binary
is always found in preference to any other installed versions.


Former-commit-id: 6f6b353c25b777caae4844db836f6546b02cc5eb