workflows/release: build macOS binaries on a macOS machine
In the future, we're going to want to sign and notarize binaries for
macOS and we can only do that from a macOS machine.  Switch the building
of macOS binaries to a macOS machine.

Note that we force matching only files when copying built assets into
bin/releases because otherwise we will copy the directory, which we
don't want.


Former-commit-id: 269a41cef742e1d48c30e8f33f3816a719bb9737