git: avoid trying to rewrite remote tags as remote branches
When invoking "git lfs migrate import --fixup", we call Git to fetch the
remote branches and tags.  However, we mark both as having the type
remote branch, so when we look up the local copies for them, we try to
use the tag name as the name of a remote tracking branch, which fails.

Instead, let's mark them as remote tags, which means we won't try to
rewrite them into branches at all.  This ensures the operation can
succeed.

While we're here, let's fix a typo and put a period at the end of our
sentence.


Former-commit-id: 2877dbe5be633ab801de3edc724bdb33de9df21a