rpm/build_rpms: remove Ruby build RPMs
When building on any CentOS version, we end up building a lot of
additional items, and this is especially true of CentOS 6, where we must
build Ruby.  These intermediate build products run us out of disk space
in CI, so let's remove the RPMs that we installed to build Ruby once
we're done with them.


Former-commit-id: 86b5bce730336133f437b77d5abaf21f7084f021