Allow literal local paths as remotes
Currently, we allow literal paths as named remotes, but not specified as
literal remotes on the command line, since they fail to validate as
URLs.  To make this experience better, let's rewrite the local paths as
file URLs using the existing endpoint code if the validation fails and
then try validation again.  We check specifically that users can use "."
as a remote since that's a common idiom to update branches.


Former-commit-id: b2d3f1d787f6b5481f6254d3bd49ba46f734b762