Allow locking/unlocking of nonexistent files
If a file is renamed or removed, a lock can be obtained and released for that file even though it doesn't exist.

Updated tests to verify that a file that doesn't exist can be locked, a file that was locked an removed can be unlocked and a file that doesn't exist can be locked and unlocked.


Former-commit-id: 88e3e1b286179136b7bcd8991c1f2525468128ad