locking: make patterns with slashes work on Windows
When we get paths from git.GetAttributePaths, the paths are in the
system native format; that is, they use the system native path
separator.  However, when creating patterns, we always want to use a
slash.  As things stand right now, lockable patterns that use a slash
never match on Windows.

Convert the patterns to use a slash so that lockable patterns work as
expected on Windows.


Former-commit-id: bf9fc6efc0b5ab6f23a693ff1b5600033845fbbb