docs/man/mangen: format output according to goimports
goimports now fails when using the -l option and any file has been
changed.  In our case, this occurs in our CI run, since we run make to
generate the commands/mancontent_gen.go file and then run make test,
which formats it.

Since the only formatting error here is an indentation problem, this is
easy to fix by inserting a tab at the beginning of the three of our
output statements.  This is both easy and straightforward, so let's do
it.


Former-commit-id: 5c0165f05a4d9c7cd657c3d16efc5e53df1770d9