git: add a function to check object ID lengths
In some places in our code, we need to check whether a potential value
has a length that could make it an object ID.  Sometimes we need to
compare exactly, and sometimes we need to compare whether it is smaller
or larger than an object ID.

Since we're now going to support SHA-256, let's add some helpers to make
this easier.  Add a slice, ObjectIDLengths, that contain all possible
hexadecimal Git object ID lengths in order so we can compare against the
first or last element.  For the case where we want to know whether a
particular length is valid, it would be burdensome to write a loop
iterating through the array, so let's add a helper function to return a
boolean and a test for that function.


Former-commit-id: be7a06b898a1e210bc96a79c169889f28138d33b