git: add a function to check for a zero object ID
There are some places in our code that we need to check whether an
object ID is the all-zeros ID.  We've had several different pieces of
code to check this, but let's put a single function together to check
this regardless of the hash algorithm in use.

Remove all the other pieces of code which check for this, including a
couple which were not even used.


Former-commit-id: 9b745d99bbdf8b52756a89e85eaeb502a31bedee