lfs: parse git cat-file --batch-check output for SHA-256
Currently, we have several hard-coded 40-based values here for the
length of an object ID.  Let's switch to finding the length of the
object ID, which will be the offset of the first space, and using that
as the length to compute other data from.  That will work for both SHA-1
and SHA-256, as well as any future hash algorithm that may be
implemented.


Former-commit-id: 95b83c5c1e96f8f321171a97973a4e9802bc8783