command/uninstall: report errors and correct usage
When uninstalling, if any "git config --remove-section" commands
failed, we would previously ignore the error and proceed without
reporting it.

We now at least report the errors in a warning message printed to
standard output, akin to the handling introduced to the
"git lfs install" command in PR #2673.

However, we likely can not alter the uninstall behaviour further
and actually exit with an error code in these cases (skipping
any attemp to uninstall our hooks) because users may depend on
the existing willingness of "git lfs uninstall" to try to
uninstall as much as possible, and may have scripts which
depend on this behaviour.  Thus we don't call os.Exit(2) in
the manner introduced to "git lfs install" in PR #3624.

On the one hand, this change means users who run an uninstall
command when no relevant Git LFS filter configurations exist
will see a "WARNING" message which may look somewhat scary.

On the other hand, we already report other errors such as
"no such file or directory" when hooks do not exist, and
by reporting these git-config(1) errors we will be able to
better support new options such as --worktree, where
git-config may report error conditions such as when the
worktreeConfig extension is not properly defined yet.

So on balance, we aim for more rather than less error
reporting.

And separately, although the usage text for all command options is
unused (because we override Cobra's usage function in run.go to
always show the man page text instead), we still correct it here
before we add any new options.


Former-commit-id: d99898dd1dfd6a28968977aab05f4c48175020d8