t: add a template directory for Git
In future versions of Git, the default branch name will likely change
from "master" to "main".  We'd like to make this change as well.  In
order to make the default branch the same in all versions of Git, let's
use a template directory to initialize the repositories that we're
creating.  This works with Git 1.8.3.1, which is available in CentOS 7,
as well as all newer versions.

For now, we set the branch name to "master", but in a future commit,
we'll change the name to "main".

We move the native_path helper into testenv.sh so that we can
canonicalize the GIT_TEMPLATE_PATH environment variable.  Git itself
canonicalizes this internally and the tests for git lfs env will fail if
we don't canonicalize it ourselves.


Former-commit-id: 3c6defef24f2e3149c14889298b0bc180651d9d6