t/t-install.sh: repair install outside repo test
The test for installing with --local outside a repository
should fail if the content of the captured error log doesn't
match the expected error string.  In practice, the captured
error log was always empty because the requireInRepo() function
prints to stdout, not stderr, and the test was capturing stderr.

The test did not fail, however, because the "set +e" setting
which allows the test to proceed past the expected failure of
the "git lfs install --local" command continued to have effect
such that the failure of the subsequent string comparison was
ignored and the test passed.

So in addition to capturing stdout rather than stderr, we also
restrict the effect of the "set +e" setting to just the
expected failure of the "git lfs install" command.  For
clarity, we rename the captured log file to out.log and do
the same in a prior test which also captures stdout as well.

We also clean up some whitespace, and remove some unneeded
global configs from the test for failed permissions when
installing with --local.


Former-commit-id: fbfd474eb19b1379eadf33842e00edd0557d4de4