t: ignore exact byte counts when pushing
One of the test objects we use in this test refers to the current branch
name and as a consequence, if we change the branch name, the number of
bytes pushed will change.  We don't really care about exactly how many
bytes were pushed, only that we pushed the right number of objects, so
shorten the string we look for to ignore the byte count.  That we upload
the right number of bytes is tested elsewhere anyway.

We could also match the entire string for a variable number of bytes
pushed, but this wouldn't be any more precise and would require the use
the -E flag to grep and escaping of the parentheses.


Former-commit-id: 5c3b94a289b269b1dea51492ff0af81d7209f386