t/t-[un]install.sh: add and refine scope checks
As the checks for local and global LFS configurations in our
install and uninstall tests of the --local option vary somewhat
in their breadth and their use of dummy text strings, we attempt
to make them more consistent and complete before adding any tests
for the new --worktree option.

To bring additional consistency we add an uninstall test
which confirms that using --local outside a repository fails,
akin to the existing installation test of that case.

And we also add matching install and uninstall tests to confirm
the expected error message when conflicting --local and --system
scope options are both supplied.


Former-commit-id: 7e779b8442fef99c47f7c0143110545cabe0b0c4