t: use repo v1 with extensions
In some of our worktree code, we use configuration options under the
extensions hierarchy.  In upstream Git, unknown extensions are rejected
if the repository format version is 1; this format version is otherwise
identical to version 0.

Until recently, Git would accept extensions with a repository format
version of 0, but fail to reject unknown extensions.  However, this
could cause serious problems with the repository, so it was recently
fixed to ignore all extensions in repository format 0.

These tests set extensions.worktreeConfig without setting the repository
format version, and consequently they fail with the latest versions of
Git.  Let's set the version appropriately so that our tests continue to
work.


Former-commit-id: 0940beb3c0e489e593e6d9ca854c0e0016b47a14