script/cibuild: make trailing whitespace check fatal
We currently intend for our CI check for trailing whitespace to be
fatal; that is, if any trailing whitespace is present, our CI script
should fail.

However, currently that doesn't work, since apparently set -e is not
concerned with a negated pipeline having a nonzero exit code.  To make
things more robust, let's make this into an if statement and exit 1 if
we match any files (that is, if grep returns 0).

Note that this used to work because it was the last command in the file,
but it no longer does for the above reason.


Former-commit-id: 5bd6350d4e75e3d48cc651920ebc7dfdd1a7338c