workflows: force Git for Windows 2.28.0
Traditionally, symbolic links on Unix have the size of the destination
in bytes, whereas symbolic links on Windows have a full block size as
their size on disk.  While this is normally not a problem, it has caused
a significant amount of pain for users that share repositories using
symlinks between a Windows version of Git and a Linux version using the
Windows Subsystem for Linux, since the index stores the size of all
files and Git would always see these files as modified.

Git for Windows 2.27.0 introduced a very desirable feature, which is
that for Git, the size of a symlink stored in the index is always
computed based on the Unix approach, which means that a repository can
be shared.  However, it came with a significant bug.

If the core.fscache option is enabled, a different lstat implementation
is used, which does not have that fix applied.  As a result, when that
option is enabled, as it is on GitHub Actions, all symbolic links
perpetually appear to be modified, even though git diff does not show
any differences.  git reset --hard similarly has no effect.  This causes
our cibuild script to fail, since we do a git reset --hard, format
files, and then look for any modified files.

Fortunately, this issue is fixed in Git for Windows 2.28.0, but that
version isn't available via Chocolatey right now, so let's roll back to
2.26.2 until GitHub Actions updates to the latest version.


Former-commit-id: 42927f3f696458be7680dc9e2b1799497985d4db