fs: gracefully handle missing temporary directory
When we are running in a repository where we lack write permissions,
we'll try to clean up temporary files nevertheless.  However, in such a
case, we may not have even been able to create a temporary directory if
one doesn't exist.

With the previous commit, we fixed the hang that occurred here, but we
print an error message like so:

  Error clearing old temp files: stat /tmp/repo/.git/lfs/tmp: no such file or directory

The user probably doesn't care about the fact that we don't have a
temporary directory in this case, and if there were no temporary files
to clean up because there was no temporary directory, then we've already
accomplished our goal.  Let's check for the temporary directory
existing, and if it doesn't, we'll just silently do nothing.


Former-commit-id: ee362cd70753eb8a0c306882a9ff0301accfa26a