Makefile: include manual pages in archives
It would be nice to include the manual pages and HTML documentation into
the archives we produce so that it's easy for users to have them without
needing to build them.  For the tar archives, this is easy: just include
"man" in the list of release targets.

It's a little more complicated for zip archives, since we need to use
the -j flag to strip off directories when we generate our archives, but
we don't want to do that here.  Instead, just explicitly add the files
to the archive with the -u (update) flag after the fact.  Note that
unlike with tar, we cannot just specify the directory, since that only
includes the directory, not its contents: we must specify the files
themselves.


Former-commit-id: d3158eb95754944414dfdcad3b0c55aabd3f1f44