script/upload: avoid using Ruby's URI.escape
Ruby's URI.escape is now deprecated because different components of a
path require different escaping rules.  However, we still want to
perform a similar escaping style, so let's do it ourselves with a
String#gsub call.  We could use CGI.escape, but that encodes spaces as
plus signs and it's unclear whether that's a desirable change to make
here.  Note that there are no negative consequences for encoding more
data than we want, since the server will decode it for us.


Former-commit-id: 765e7049b102ce75b7591fcaba660a12a087de94