tools: add a function to canonicalize paths
We have several different places in our code that need to canonicalize
paths.  In our case, that usually involves resolving a Cygwin path to a
native Windows path, turning the path absolute, and calling
filepath.EvalSymlinks.  Let's add a function that does exactly that and
call it from the places in the git package where we do this already.

We pass an additional argument to indicate whether it's acceptable if
the path is missing, and if so, we return an absolute but
uncanonicalized path in that case.  This is useful for canonicalizing
Git environment variables which may or may not point to a valid
location; we want Git, not us, to make the decision about whether a
missing path is a problem in such a case.


Former-commit-id: 19b2cd8e909bd6211b13686b281cb418befbc767