Add --above parameter to 'migrate import'.
The purpose is to allow users to tell git-lfs to migrate
all files over a particular threshold without having to determine
the various paths and patterns that would. This is very useful
for large code bases with inconsistent history of adding large
files in different places. This option makes it possible to
supply `git-lfs migrate import --everything --above=1Mb` and
have git-lfs do the convenient thing and rewrite history.


Former-commit-id: 5a4f2eb7f9761f36f9cb6cd24bb61fe8deb63050