Drop support for CentOS 6
CentOS 6 went out of support on November 30, 2020.  Moreover, building
packages for it requires that we build Ruby in addition to all the
gems, plus a version of Git, meaning we spend significant CI resources
and time on this platform.

Since there's no longer any free security support and we cannot
reasonably be expected to run insecure code in CI, since that's
dangerous to others on the Internet, let's drop support for CentOS 6.


Former-commit-id: d69e28de4a5d401a4255a12152f2332af79eee83